fix(ext/console): fix inspecting iterators error. (#20720)
Fixes #19776 and #20676.
